LogInViewCell deleted/Code moved to LogInViewController: ScrollView added/Keyboard tuned
